GCW-3169 refactored messages-screen
